+++ packages/address_autocomplete_util/lib/address_result_model.dart
@@ -1,5 +1,7 @@
 import 'dart:collection';
 
+import 'package:address_autocomplete_util/enums.dart';
+
 /// {@template AddressResults}
 /// Extracts the address string from Google Places API response.
 ///
@@ -8,13 +10,19 @@ import 'dart:collection';
 /// {@endtemplate}
 class AddressResults extends ListMixin<String> {
   /// {@macro AddressResults}
-  AddressResults({
+  AddressResults._privateConstructor({
     List<String>? predictions,
+    required this.requestStatus,
   }) : this.predictions = predictions ??= [];
 
   List<String> predictions;
+  PlacesStatusCodes requestStatus;
 
   /// {@macro AddressResults}
+  ///
+  /// See [Place Autocomplete
+  /// response](https://developers.google.com/maps/documentation/places/web-service/autocomplete?hl=en_US#place_autocomplete_responses)
+  /// for full json output.
   factory AddressResults.fromJson(Map<String, dynamic> json) {
     List<Map<String, dynamic>> predictionsMap =
         (json["predictions"] as List).cast<Map<String, dynamic>>();
@@ -23,7 +31,14 @@ class AddressResults extends ListMixin<String> {
             predictionJson['description'] as String)
         .toList();
 
-    return AddressResults(predictions: _predictions);
+    String statusCode = json['status'];
+    PlacesStatusCodes requestStatus = PlacesStatusCodes.values.singleWhere(
+        (placesStatusCode) => placesStatusCode.value == statusCode);
+
+    return AddressResults._privateConstructor(
+      predictions: _predictions,
+      requestStatus: requestStatus,
+    );
   }
 
   @override