+++ packages/address_autocomplete_util/lib/places_exceptions.dart
@@ -1,0 +1,11 @@
+class PlacesException implements Exception {
+  final dynamic message;
+
+  PlacesException([this.message]);
+
+  String toString() {
+    Object? message = this.message;
+    if (message == null) return "PlacesException";
+    return "PlacesException: $message";
+  }
+}