+++ packages/address_autocomplete_util/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
+### Added
+- `PlacesException` and `PlacesStatusCode◊s` to provide additional context on
+  failed requests (e.g. invalid APIKey)
+- Updated example with try/catch on `AddressFinder.fetchAddress()`.
 
 ## [0.0.1] - 2021-07-19
 Initial release.